Initial Commit.
Previous theme was a a StrapPress child theme with many non-standard mods.
This theme was made as a child them against `_s`. They recommend direct
devlopment (not child), but I chose to keep it decoupled because I wasn't
sold on targeting some of their decisions in the long run.

Initial ad hoc work on the CSS was migrated to PostCSS with Tailwind, to
enable more coherent responsive design and graceful degradation.

This commit is the version that ran on the web for some time that targets
the original messy metadata made with a custom post type plugin. The site
metadata was later migrated to a more coherent format, that required a
more advanced archive page processing.