`v2.0-dev` colorized theme.
At the request of the patrons, I shifted some focus to adding more color. This
had been discussed earlier on, but was brought forward in the timeline.

I used CSS var tactics to create a pallete based summary of the effects already
in play, plus some new ones I had discovered along the way.

I also did cleanup as I moved through the site sections.

This commit mostly represents the merge in of the end of the last work session, which left open a couple layout issues.

For logistical reasons, this commit was merged in without a live test, and
was further revisioned, for cleanup and hopeful bug fixes.

Fixes to bring the site live from this commit should be expected.